SYM-7534: Moved ApplicationHealthTracker to symmetric-core, added static holder methods, and made the URI handler check isAlive()
